use the correct variable type, to save a little memory in the alignment_t struct too

+++ src/align/include/align_parameters.hpp
@@ -22,7 +22,7 @@ struct Parameters {
     //bool exact_wfa;                               // use exact WFA, avoiding adaptive wavefront reduction
 
     //wflambda
-    uint32_t wflambda_segment_length;             //segment length for wflambda
+    uint16_t wflambda_segment_length;             //segment length for wflambda
     uint32_t wflambda_min_wavefront_length;       //wavefront length to trigger reduction (how wide should it be)
     uint32_t wflambda_max_distance_threshold;     //maximum distance (in WFA diagonals) that a wavefront can fall behind the furthest
